+++ scenario_runner.py
@@ -27,6 +27,7 @@ import os
 import signal
 import sys
 import time
+import json
 import pkg_resources
 
 import carla
@@ -40,7 +41,7 @@ from srunner.tools.scenario_parser import ScenarioConfigurationParser
 from srunner.tools.route_parser import RouteParser
 
 # Version of scenario_runner
-VERSION = 0.6
+VERSION = '0.9.9'
 
 
 class ScenarioRunner(object):
@@ -175,7 +176,6 @@ class ScenarioRunner(object):
             settings.fixed_delta_seconds = None
             self.world.apply_settings(settings)
 
-        self.client.stop_recorder()
         self.manager.cleanup()
 
         CarlaDataProvider.cleanup()
@@ -255,6 +255,38 @@ class ScenarioRunner(object):
             if not (self._args.output or filename or junit_filename):
                 print("Please run with --output for further information")
 
+    def _record_criteria(self, criteria, name):
+        """
+        Filter the JSON serializable attributes of the criterias and
+        dumps them into a file. This will be used by the metrics manager,
+        in case the user wants specific information about the criterias.
+        """
+        file_name = name[:-4] + ".json"
+
+        # Filter the attributes that aren't JSON serializable
+        with open('temp.json', 'w') as fp:
+
+            criteria_dict = {}
+            for criterion in criteria:
+
+                criterion_dict = criterion.__dict__
+                criteria_dict[criterion.name] = {}
+
+                for key in criterion_dict:
+                    if key != "name":
+                        try:
+                            key_dict = {key: criterion_dict[key]}
+                            json.dump(key_dict, fp, sort_keys=False, indent=4)
+                            criteria_dict[criterion.name].update(key_dict)
+                        except TypeError:
+                            pass
+
+        os.remove('temp.json')
+
+        # Save the criteria dictionary into a .json file
+        with open(file_name, 'w') as fp:
+            json.dump(criteria_dict, fp, sort_keys=False, indent=4)
+
     def _load_and_wait_for_world(self, town, ego_vehicles=None):
         """
         Load a new CARLA world and provide data to CarlaDataProvider
@@ -289,6 +321,7 @@ class ScenarioRunner(object):
 
         CarlaDataProvider.set_client(self.client)
         CarlaDataProvider.set_world(self.world)
+        CarlaDataProvider.set_traffic_manager_port(int(self._args.trafficManagerPort))
 
         # Wait for the world to be ready
         if CarlaDataProvider.is_sync_mode():
@@ -351,17 +384,23 @@ class ScenarioRunner(object):
             return False
 
         try:
-            # Load scenario and run it
             if self._args.record:
-                self.client.start_recorder("{}/{}.log".format(os.getenv('SCENARIO_RUNNER_ROOT', "./"), config.name))
+                recorder_name = "{}/{}/{}.log".format(
+                    os.getenv('SCENARIO_RUNNER_ROOT', "./"), self._args.record, config.name)
+                self.client.start_recorder(recorder_name, True)
+
+            # Load scenario and run it
             self.manager.load_scenario(scenario, self.agent_instance)
             self.manager.run_scenario()
 
             # Provide outputs if required
             self._analyze_scenario(config)
 
-            # Remove all actors
+            # Remove all actors, stop the recorder and save all criterias (if needed)
             scenario.remove_all_actors()
+            if self._args.record:
+                self.client.stop_recorder()
+                self._record_criteria(self.manager.scenario.get_criteria(), recorder_name)
 
             result = True
 
@@ -456,43 +495,50 @@ def main():
     main function
     """
     description = ("CARLA Scenario Runner: Setup, Run and Evaluate scenarios using CARLA\n"
-                   "Current version: " + str(VERSION))
+                   "Current version: " + VERSION)
 
+    # pylint: disable=line-too-long
     parser = argparse.ArgumentParser(description=description,
                                      formatter_class=RawTextHelpFormatter)
+    parser.add_argument('-v', '--version', action='version', version='%(prog)s ' + VERSION)
     parser.add_argument('--host', default='127.0.0.1',
                         help='IP of the host server (default: localhost)')
     parser.add_argument('--port', default='2000',
                         help='TCP port to listen to (default: 2000)')
+    parser.add_argument('--timeout', default="10.0",
+                        help='Set the CARLA client timeout value in seconds')
+    parser.add_argument('--trafficManagerPort', default='8000',
+                        help='Port to use for the TrafficManager (default: 8000)')
     parser.add_argument('--sync', action='store_true',
                         help='Forces the simulation to run synchronously')
-    parser.add_argument('--debug', action="store_true", help='Run with debug output')
+    parser.add_argument('--list', action="store_true", help='List all supported scenarios and exit')
+
+    parser.add_argument(
+        '--scenario', help='Name of the scenario to be executed. Use the preposition \'group:\' to run all scenarios of one class, e.g. ControlLoss or FollowLeadingVehicle')
+    parser.add_argument('--openscenario', help='Provide an OpenSCENARIO definition')
+    parser.add_argument(
+        '--route', help='Run a route as a scenario (input: (route_file,scenario_file,[route id]))', nargs='+', type=str)
+
+    parser.add_argument(
+        '--agent', help="Agent used to execute the scenario. Currently only compatible with route-based scenarios.")
+    parser.add_argument('--agentConfig', type=str, help="Path to Agent's configuration file", default="")
+
     parser.add_argument('--output', action="store_true", help='Provide results on stdout')
     parser.add_argument('--file', action="store_true", help='Write results into a txt file')
     parser.add_argument('--junit', action="store_true", help='Write results into a junit file')
     parser.add_argument('--outputDir', default='', help='Directory for output files (default: this directory)')
-    parser.add_argument('--waitForEgo', action="store_true", help='Connect the scenario to an existing ego vehicle')
+
     parser.add_argument('--configFile', default='', help='Provide an additional scenario configuration file (*.xml)')
     parser.add_argument('--additionalScenario', default='', help='Provide additional scenario implementations (*.py)')
+
+    parser.add_argument('--debug', action="store_true", help='Run with debug output')
     parser.add_argument('--reloadWorld', action="store_true",
                         help='Reload the CARLA world before starting a scenario (default=True)')
-    # pylint: disable=line-too-long
-    parser.add_argument(
-        '--scenario', help='Name of the scenario to be executed. Use the preposition \'group:\' to run all scenarios of one class, e.g. ControlLoss or FollowLeadingVehicle')
+    parser.add_argument('--record', type=str, default='',
+                        help='Path were the files will be saved, relative to SCENARIO_RUNNER_ROOT.\nActivates the CARLA recording feature and saves to file all the criteria information.')
     parser.add_argument('--randomize', action="store_true", help='Scenario parameters are randomized')
     parser.add_argument('--repetitions', default=1, type=int, help='Number of scenario executions')
-    parser.add_argument('--list', action="store_true", help='List all supported scenarios and exit')
-    parser.add_argument(
-        '--agent', help="Agent used to execute the scenario (optional). Currently only compatible with route-based scenarios.")
-    parser.add_argument('--agentConfig', type=str, help="Path to Agent's configuration file", default="")
-    parser.add_argument('--openscenario', help='Provide an OpenSCENARIO definition')
-    parser.add_argument(
-        '--route', help='Run a route as a scenario (input: (route_file,scenario_file,[number of route]))', nargs='+', type=str)
-    parser.add_argument('--record', action="store_true",
-                        help='Use CARLA recording feature to create a recording of the scenario')
-    parser.add_argument('--timeout', default="10.0",
-                        help='Set the CARLA client timeout value in seconds')
-    parser.add_argument('-v', '--version', action='version', version='%(prog)s ' + str(VERSION))
+    parser.add_argument('--waitForEgo', action="store_true", help='Connect the scenario to an existing ego vehicle')
 
     arguments = parser.parse_args()
     # pylint: enable=line-too-long