+++ Docs/CHANGELOG.md
@@ -25,12 +25,13 @@
 * The new weather parameters (related to fog) are now correctly read when running scenarios outside routes.
 * Enable weather animation during scenario execution (requires ephem pip package)
 * Changed manual control to be in par with the CARLA version. Among others, added vehicle lights, recording and some new sensors
-* Removed unsupported scenarios (ChallengeBasic and BackgroundActivity, VehicleTurnLeftAtJunction) 
+* Removed unsupported scenarios (ChallengeBasic and BackgroundActivity, VehicleTurnLeftAtJunction)
+* Added a new metrics module, which gives access to all the information about a scenario in order to allow the user to extract any desired information about the simulation. More information [here](metrics_module.md)
 * OpenSCENARIO support:
     - Added support for controllers and provided default implementations for vehicles and pedestrians. This required changing the handling of actors, which results in that now all actors are controlled by an OSC controller. Supported controllers:
         - Pedestrian controller
         - NPC vehicle controller (based on CARLA LocalPlanner)
-        - Simple vehicle controller to set velocities not brake/throttle
+        - Simple vehicle controller to set velocities not brake/throttle, and consider obstacles in the forward-facing region.
         - External controller (to forward control to external entities)
     - Added initial speed support for pedestrians for OpenSCENARIO
     - Support for EnvironmentActions within Story (before only within Init). This allows changing weather conditions during scenario execution
@@ -42,15 +43,21 @@
     - Added support for EndOfRoadCondition
     - Added support for TimeHeadwayCondition
     - Added support for TrafficSignalCondition
+    - Added support for AcquirePositionAction
     - Extended FollowLeadingVehicle example to illustrate weather changes
     - Created example scenarios to illustrate usage of controllers and weather changes
     - Reworked the handling of Catalogs to make it compliant to the 1.0 version (relative paths have to be relative to the scenario file)
     - The RoadNetwork can be defined as global Parameter
     - Fixed handling of relative positions with negative offset
     - Added support for local ParamaterDeclarations
+    - Added support for Parameters within Catalogs
+    - Added support for ParameterAssignments for CatalogReferences
+    - Fixed name handling of Parameters: Parameter declerations must not start with a leading '$', but when the parameter is used a leading '$' is required.
+    - Fixed use of Parameters for multiple instances of the same Catalog element
     - Fixed use of relative initial positions for any actor
     - Added possibility to use synchronous execution mode with OpenSCENARIO
     - Fixed use of relative paths in CustomCommandAction
+    - Fixed use of ControllerCatalogs
 * Atomics:
     - Several new atomics to enable usage of OSC controllers
     - WeatherBehavior to simulate weather over time
@@ -68,6 +75,9 @@
     - Changed the inputs to TrafficLightStateSetter to match the other atomics, but the functionality remains unchanged
 
 ### :bug: Bug Fixes
+* Support OpenSCENARIO parameters also if they're only part of a string value
+* Support Routes in Catalogs
+* Fix parsing of properties within ControllerCatalogs
 * Add cleanup of instantiated OpenSCENARIO controllers
 * Do not register SIGHUP signal in windows
 * Fixed initial speed of vehicles using OpenSCENARIO
@@ -82,7 +92,11 @@
 * Fixed bug causing NoSignalJunctionCrossing to not output the results of the scenario
 * Fixed bug causing SyncArrival to fail when the actor was destroyed after the behavior ended
 * Fixed bug with ending roads near stop signals to break the simulation
-
+* Fixed exception bug in spawn function of CarlaDataProvider
+* Fixed access to private member of CARLA LocalPlanner inside OSC NpcVehicleControl
+* Fixed handling of OSC LanePosition (#625)
+### :ghost: Maintenance
+* Exposed traffic manager port flag to enable the execution of multiple scenarios on a single machine.
 
 ## CARLA ScenarioRunner 0.9.9
 ### :rocket: New Features